Extract seller detail cards and sheets into their own components
The seller detail route had grown to 1075 lines holding eleven
components. The read-only cards and the three edit sheets now live
beside the other seller components; the route keeps the lifecycle
actions, which belong to the page.

No behaviour change.
